requirements.txt
pdfplumber
spacy
pyyaml